requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
Deprecated==1.2.13
environ==1.0
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pynubank==2.21.1
pyOpenSSL==23.0.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.28.2
requests-pkcs12==1.15
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
python-dotenv==1.0.0
